Combine generateKey operations of ECDSA and ECDH
