+++ components/script/dom/webcrypto/subtlecrypto/ecdh_operation.rs
@@ -2,10 +2,9 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at https://mozilla.org/MPL/2.0/. */
 
+use elliptic_curve::Curve;
 use elliptic_curve::generic_array::typenum::Unsigned;
-use elliptic_curve::rand_core::OsRng;
 use elliptic_curve::sec1::{FromEncodedPoint, ToEncodedPoint, ValidatePublicKey};
-use elliptic_curve::{Curve, SecretKey};
 use js::context::JSContext;
 use p256::NistP256;
 use p384::NistP384;
@@ -25,6 +24,7 @@ use crate::dom::bindings::root::DomRoot;
 use crate::dom::bindings::str::DOMString;
 use crate::dom::cryptokey::{CryptoKey, Handle};
 use crate::dom::globalscope::GlobalScope;
+use crate::dom::subtlecrypto::ec_common::EcAlgorithm;
 use crate::dom::subtlecrypto::{
     CryptoAlgorithm, ExportedKey, JsonWebKeyExt, JwkStringField, KeyAlgorithmAndDerivatives,
     NAMED_CURVE_P256, NAMED_CURVE_P384, NAMED_CURVE_P521, SUPPORTED_CURVES, SubtleEcKeyAlgorithm,
@@ -39,115 +39,14 @@ pub(crate) fn generate_key(
     extractable: bool,
     usages: Vec<KeyUsage>,
 ) -> Result<CryptoKeyPair, Error> {
-    // Step 1. If usages contains an entry which is not "deriveKey" or "deriveBits" then throw a
-    // SyntaxError.
-    if usages
-        .iter()
-        .any(|usage| !matches!(usage, KeyUsage::DeriveKey | KeyUsage::DeriveBits))
-    {
-        return Err(Error::Syntax(Some(
-            "Usages contains an entry which is not \"deriveKey\" or \"deriveBits\"".to_string(),
-        )));
-    }
-
-    // Step 2.
-    // If the namedCurve member of normalizedAlgorithm is "P-256", "P-384" or "P-521":
-    //     Generate an Elliptic Curve key pair, as defined in [RFC6090] with domain parameters for
-    //     the curve identified by the namedCurve member of normalizedAlgorithm.
-    // If the namedCurve member of normalizedAlgorithm is a value specified in an applicable
-    // specification that specifies the use of that value with ECDH:
-    //     Perform the ECDH generation steps specified in that specification, passing in
-    //     normalizedAlgorithm and resulting in an elliptic curve key pair.
-    // Otherwise:
-    //     throw a NotSupportedError
-    // Step 3. If performing the operation results in an error, then throw a OperationError.
-    let (private_key_handle, public_key_handle) = match normalized_algorithm.named_curve.as_str() {
-        NAMED_CURVE_P256 => {
-            let private_key = SecretKey::<NistP256>::random(&mut OsRng);
-            let public_key = private_key.public_key();
-            (
-                Handle::P256PrivateKey(private_key),
-                Handle::P256PublicKey(public_key),
-            )
-        },
-        NAMED_CURVE_P384 => {
-            let private_key = SecretKey::<NistP384>::random(&mut OsRng);
-            let public_key = private_key.public_key();
-            (
-                Handle::P384PrivateKey(private_key),
-                Handle::P384PublicKey(public_key),
-            )
-        },
-        NAMED_CURVE_P521 => {
-            let private_key = SecretKey::<NistP521>::random(&mut OsRng);
-            let public_key = private_key.public_key();
-            (
-                Handle::P521PrivateKey(private_key),
-                Handle::P521PublicKey(public_key),
-            )
-        },
-        _ => {
-            return Err(Error::NotSupported(Some(format!(
-                "Unsupported namedCurve: {}",
-                normalized_algorithm.named_curve
-            ))));
-        },
-    };
-
-    // Step 4. Let algorithm be a new EcKeyAlgorithm object.
-    // Step 5. Set the name member of algorithm to "ECDH".
-    // Step 6. Set the namedCurve attribute of algorithm to equal the namedCurve member of
-    // normalizedAlgorithm.
-    let algorithm = SubtleEcKeyAlgorithm {
-        name: CryptoAlgorithm::Ecdh,
-        named_curve: normalized_algorithm.named_curve.clone(),
-    };
-
-    // Step 7. Let publicKey be a new CryptoKey representing the public key of the generated key pair.
-    // Step 8. Set the [[type]] internal slot of publicKey to "public"
-    // Step 9. Set the [[algorithm]] internal slot of publicKey to algorithm.
-    // Step 10. Set the [[extractable]] internal slot of publicKey to true.
-    // Step 11. Set the [[usages]] internal slot of publicKey to be the empty list.
-    let public_key = CryptoKey::new(
+    ec_common::generate_key(
+        EcAlgorithm::Ecdh,
         cx,
         global,
-        KeyType::Public,
-        true,
-        KeyAlgorithmAndDerivatives::EcKeyAlgorithm(algorithm.clone()),
-        Vec::new(),
-        public_key_handle,
-    );
-
-    // Step 12. Let privateKey be a new CryptoKey representing the private key of the generated key pair.
-    // Step 13. Set the [[type]] internal slot of privateKey to "private"
-    // Step 14. Set the [[algorithm]] internal slot of privateKey to algorithm.
-    // Step 15. Set the [[extractable]] internal slot of privateKey to extractable.
-    // Step 16. Set the [[usages]] internal slot of privateKey to be the usage intersection of
-    // usages and [ "deriveKey", "deriveBits" ].
-    let private_key = CryptoKey::new(
-        cx,
-        global,
-        KeyType::Private,
+        normalized_algorithm,
         extractable,
-        KeyAlgorithmAndDerivatives::EcKeyAlgorithm(algorithm),
-        usages
-            .iter()
-            .filter(|usage| matches!(usage, KeyUsage::DeriveKey | KeyUsage::DeriveBits))
-            .cloned()
-            .collect(),
-        private_key_handle,
-    );
-
-    // Step 17. Let result be a new CryptoKeyPair dictionary.
-    // Step 18. Set the publicKey attribute of result to be publicKey.
-    // Step 19. Set the privateKey attribute of result to be privateKey.
-    let result = CryptoKeyPair {
-        publicKey: Some(public_key),
-        privateKey: Some(private_key),
-    };
-
-    // Step 20. Return result.
-    Ok(result)
+        usages,
+    )
 }
 
 /// <https://w3c.github.io/webcrypto/#ecdh-operations-derive-bits>
